Complete Phase 1: Code simplification and refactoring. Pruned dead code, consolidated core logic into fast_mypyc/engine.py and fast_numba/*, added type hints, and documented percentile algorithm. Further optimization and finalization phases pending.
